+++ inselect/gui/views/summary.py
@@ -1,5 +1,7 @@
 import locale
-from PySide.QtGui import QAbstractItemView, QHBoxLayout, QLabel, QWidget
+from PySide import QtCore
+from PySide.QtGui import (QAbstractItemView, QHBoxLayout, QLabel, QSizePolicy,
+                          QWidget)
 
 from inselect.lib.utils import debug_print
 
@@ -11,36 +13,39 @@ class SummaryView(QAbstractItemView):
         # This view is not visible
         super(SummaryView, self).__init__(None)
 
-        layout = QHBoxLayout()
-
-        self.n_boxes = QLabel()
-        layout.addWidget(self.n_boxes)
+        self.info = QLabel()
 
         # Last item has stretch greater than zero to force all labels to be
         # left-aligned
-        self.n_selected = QLabel()
-        layout.addWidget(self.n_selected, stretch=1)
+        layout = QHBoxLayout()
+
+        # Smaller margins than the defaults
+        layout.setContentsMargins(
+            8,  # left
+            2,  # top
+            0,  # right
+            2   # bottom
+        )
+
+        layout.addWidget(self.info, stretch=1)
 
         self.widget = QWidget(parent)
         self.widget.setLayout(layout)
 
-    def _n_boxes(self, n):
-        self.n_boxes.setText(
-            '{0} boxes'.format(locale.format("%d", n, grouping=True))
-        )
-
-    def _n_selected(self, n):
-        self.n_selected.setText(
-            '{0} selected'.format(locale.format("%d", n, grouping=True))
-        )
+    def _updated(self, n, selected):
+        template = u'{0} boxes / {1} selected / {2}'
+        self.info.setText(template.format(
+            locale.format("%d", n, grouping=True),
+            locale.format("%d", len(selected), grouping=True),
+            selected[0].data(QtCore.Qt.DisplayRole) if 1 == len(selected) else ''
+        ))
 
     def reset(self):
         """QAbstractItemView virtual
         """
         debug_print('SummaryView.reset')
         super(SummaryView, self).reset()
-        self._n_boxes(self.model().rowCount())
-        self._n_selected(0)
+        self._updated(self.model().rowCount(), [])
 
     def setModel(self, model):
         """QAbstractItemView virtual
@@ -53,16 +58,23 @@ class SummaryView(QAbstractItemView):
         """QAbstractItemView virtual
         """
         debug_print('SummaryView.dataChanged')
-        self._n_boxes(self.model().rowCount())
+        self._updated(
+            self.model().rowCount(), self.selectionModel().selectedIndexes()
+        )
 
     def selectionChanged(self, selected, deselected):
         """QAbstractItemView slot
         """
         debug_print('SummaryView.selectionChanged')
-        self._n_selected(len(self.selectionModel().selectedIndexes()))
+        self._updated(
+            self.model().rowCount(), self.selectionModel().selectedIndexes()
+        )
 
     def rowsAboutToBeRemoved(self, parent, start, end):
         """QAbstractItemView slot
         """
         debug_print('SummaryView.rowsAboutToBeRemoved')
-        self._n_boxes(self.model().rowCount() - (end - start))
+        self._updated(
+            self.model().rowCount() - (end - start),
+            self.selectionModel().selectedIndexes()
+        )