+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Version 0.1.26
 -------------
 - Fixed #256 - Disable template 'Reload' command when default template selected
 - Fixed #254 - An Error Occurred: {'Choices with data ...
+- Fixed #253 - Numbered bounding boxes
+- Fixed #218 - N boxes / N selected widget to a status bar
 
 Version 0.1.25
 -------------